Moving block definitions
Moving block definitions that could be shared between games into appengine/js with the common theme of the blocks having a look more similar to javascript.
