+++ appengine/pond/duck/template.soy
@@ -68,26 +68,26 @@
     <category name="Pond">
       <block type="pond_cannon">
         <value name="DEGREE">
-          <shadow type="pond_math_number">
+          <shadow type="js_math_number">
             <field name="NUM">0</field>
           </shadow>
         </value>
         <value name="RANGE">
-          <shadow type="pond_math_number">
+          <shadow type="js_math_number">
             <field name="NUM">70</field>
           </shadow>
         </value>
       </block>
       <block type="pond_scan">
         <value name="DEGREE">
-          <shadow type="pond_math_number">
+          <shadow type="js_math_number">
             <field name="NUM">0</field>
           </shadow>
         </value>
       </block>
       <block type="pond_swim">
         <value name="DEGREE">
-          <shadow type="pond_math_number">
+          <shadow type="js_math_number">
             <field name="NUM">0</field>
           </shadow>
         </value>
@@ -99,17 +99,17 @@
       <block type="pond_health"></block>
     </category>
     <category name="Logic">
-      <block type="pond_controls_if"></block>
+      <block type="js_controls_if"></block>
       <block type="logic_compare"></block>
       <block type="logic_operation"></block>
       <block type="logic_boolean"></block>
     </category>
     <category name="Loops">
-      <block type="pond_loops_while"></block>
+      <block type="js_loops_while"></block>
     </category>
     <category name="Math">
-      <block type="pond_math_number"></block>
-      <block type="pond_math_arithmetic">
+      <block type="js_math_number"></block>
+      <block type="js_math_arithmetic">
         <value name="A">
           <shadow type="math_number">
             <field name="NUM">1</field>
@@ -121,16 +121,16 @@
           </shadow>
         </value>
       </block>
-      <block type="pond_math_single">
+      <block type="js_math_single">
         <value name="NUM">
           <shadow type="math_number">
             <field name="NUM">9</field>
           </shadow>
         </value>
       </block>
-      <block type="pond_math_change">
+      <block type="js_math_change">
         <value name="DELTA">
-          <shadow type="pond_math_number">
+          <shadow type="js_math_number">
             <field name="NUM">1</field>
           </shadow>
         </value>
